+++ User/app/src/main/java/com/example/jub_jub_android/ui/adapter/recyclerview/ManageItem_RecyclerViewAdpater.kt
@@ -1,20 +1,19 @@
-package com.example.jub_jub_android.ui.adapter
+package com.example.jub_jub_android.ui.adapter.recyclerview
 
 import android.content.Intent
+import android.content.Intent.FLAG_ACTIVITY_NEW_TASK
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.example.jub_jub_android.R
 import com.example.jub_jub_android.entity.dataclass.ItemStatus
-import com.example.jub_jub_android.entity.dataclass.LaptopStatus
 import com.example.jub_jub_android.ui.activity.ModifyItemActivity
-import com.example.jub_jub_android.ui.activity.ModifyLaptopActivity
 import com.example.jub_jub_android.ui.util.MyUtil
 import kotlinx.android.synthetic.main.layout_equipmentlist_item.view.*
 import java.util.ArrayList
 
-class ManageLaptop_RecyclerViewAdapter(var dataList: ArrayList<LaptopStatus>): RecyclerView.Adapter<ManageLaptop_RecyclerViewAdapter.ViewHolder>() {
+class ManageItem_RecyclerViewAdpater(var dataList: ArrayList<ItemStatus>): RecyclerView.Adapter<ManageItem_RecyclerViewAdpater.ViewHolder>() {
 
     //private var devideditemStatusList = dataList
 
@@ -34,27 +33,27 @@ class ManageLaptop_RecyclerViewAdapter(var dataList: ArrayList<LaptopStatus>): R
 
 
     class ViewHolder(view: View): RecyclerView.ViewHolder(view) {
-        fun bindItemStatusListItems(data: LaptopStatus){
+        fun bindItemStatusListItems(data: ItemStatus){
 
             setTextViewsText(data)
 
             //각각의 아이템 클릭시
             itemView.setOnClickListener {
 
-                var intent = Intent(itemView.context, ModifyLaptopActivity::class.java)
+                var intent = Intent(itemView.context, ModifyItemActivity::class.java)
 
                 intent.putExtra("Data", data)
-                intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
+                intent.addFlags(FLAG_ACTIVITY_NEW_TASK)
 
                 itemView.context.startActivity(intent)
             }
         }
 
-        private fun setTextViewsText(data: LaptopStatus) {
-            itemView.imageView_ItemImage_Item.setImageBitmap(MyUtil().convertBase64ToBitmap(data.image))
-            itemView.textView_ItemName_Item.text = data.name
-            itemView.textView_ItemCategory_Item.text = data.category
-            itemView.textView_ItemCount_Item.text = "수량 : ${data.count}개"
+        private fun setTextViewsText(data: ItemStatus) {
+            itemView.imageView_ItemImage.setImageBitmap(MyUtil().convertBase64ToBitmap(data.image))
+            itemView.textView_ItemName.text = data.name
+            itemView.textView_ItemCategory.text = data.category
+            itemView.textView_ItemCount.text = "수량 : ${data.count}개"
 
         }
