+++ User/app/src/main/java/com/example/jub_jub_android/ui/adapter/recyclerview/ManageLaptop_RecyclerViewAdapter.kt
@@ -1,19 +1,18 @@
-package com.example.jub_jub_android.ui.adapter
+package com.example.jub_jub_android.ui.adapter.recyclerview
 
 import android.content.Intent
-import android.content.Intent.FLAG_ACTIVITY_NEW_TASK
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.example.jub_jub_android.R
-import com.example.jub_jub_android.entity.dataclass.ItemStatus
-import com.example.jub_jub_android.ui.activity.ModifyItemActivity
+import com.example.jub_jub_android.entity.dataclass.LaptopStatus
+import com.example.jub_jub_android.ui.activity.ModifyLaptopActivity
 import com.example.jub_jub_android.ui.util.MyUtil
 import kotlinx.android.synthetic.main.layout_equipmentlist_item.view.*
 import java.util.ArrayList
 
-class ManageItem_RecyclerViewAdpater(var dataList: ArrayList<ItemStatus>): RecyclerView.Adapter<ManageItem_RecyclerViewAdpater.ViewHolder>() {
+class ManageLaptop_RecyclerViewAdapter(var dataList: ArrayList<LaptopStatus>): RecyclerView.Adapter<ManageLaptop_RecyclerViewAdapter.ViewHolder>() {
 
     //private var devideditemStatusList = dataList
 
@@ -33,27 +32,27 @@ class ManageItem_RecyclerViewAdpater(var dataList: ArrayList<ItemStatus>): Recyc
 
 
     class ViewHolder(view: View): RecyclerView.ViewHolder(view) {
-        fun bindItemStatusListItems(data: ItemStatus){
+        fun bindItemStatusListItems(data: LaptopStatus){
 
             setTextViewsText(data)
 
             //각각의 아이템 클릭시
             itemView.setOnClickListener {
 
-                var intent = Intent(itemView.context, ModifyItemActivity::class.java)
+                var intent = Intent(itemView.context, ModifyLaptopActivity::class.java)
 
                 intent.putExtra("Data", data)
-                intent.addFlags(FLAG_ACTIVITY_NEW_TASK)
+                intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
 
                 itemView.context.startActivity(intent)
             }
         }
 
-        private fun setTextViewsText(data: ItemStatus) {
-            itemView.imageView_ItemImage_Item.setImageBitmap(MyUtil().convertBase64ToBitmap(data.image))
-            itemView.textView_ItemName_Item.text = data.name
-            itemView.textView_ItemCategory_Item.text = data.category
-            itemView.textView_ItemCount_Item.text = "수량 : ${data.count}개"
+        private fun setTextViewsText(data: LaptopStatus) {
+            itemView.imageView_ItemImage.setImageBitmap(MyUtil().convertBase64ToBitmap(data.image))
+            itemView.textView_ItemName.text = data.name
+            itemView.textView_ItemCategory.text = data.category
+            itemView.textView_ItemCount.text = "수량 : ${data.count}개"
 
         }
